requirements.txt
pandas
transformers
requests>=2.25.0
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
openai
asyncpg
asyncio
aiohttp
nltk
rouge-score
bert-score